refactor(drilldemon): Modularize Drill Demon random event - Added Drill Demon config option - Refactored Drill Demon code into its own module 	- Added DrillDemonHelper and DrillDemonOverlay 	- Refactored DrillExercise enum into its own class
